kmit-radiology-main-3/requirements.txt
cycler==0.10.0
decorator==4.4.2
Django==2.2.7
django-crispy-forms==1.11.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.4.1
networkx==2.5.1
nibabel==3.2.1
numpy==1.20.2
opencv-python==4.5.1.48
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
pydicom==2.1.2
pyparsing==2.4.7
pypng==0.0.20
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
scikit-image==0.18.1
scipy==1.6.3
six==1.15.0
sqlparse==0.4.1
tifffile==2021.4.8
tqdm==4.60.0
